removed dodge as a skill and instead made it a feat, deftness.  Deftness can be mset until a point system is established and built. Advise to cset dodge_mod to at least 3.  Also manually edit skills.dat to remove dodge from the skill list.
when fight calls to check for dodge, instead of looking at the gsn_dodge for the chance variable, it now uses the equation
chances = ((25 * ch->pcdata->deftness) / sysdata.dodge_mod)

mset allows for a value max of 4 to be set on deftness. New characters are created with this value set at 0. Score display updated
to show deftness and level as well.
